embedding/requirements.txt
scipy
tensorflow
tensorflow_hub
tf-models-official==2.3.0
